api/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0
git+git://github.com/arteria/django-background-tasks.git@12c8f328e4ba704bd7d91534bb6569e70197b949 # django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.0
django-oauth-toolkit==1.2.0
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.8
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
